requirements.txt
autogen
streamlit
groq
openai
anthropic
ollama
python-dotenv 